+++ src/main/kotlin/dev/turingcomplete/intellijdevelopertoolsplugin/DeveloperToolConfiguration.kt
@@ -7,12 +7,12 @@ import dev.turingcomplete.intellijdevelopertoolsplugin.DeveloperToolConfiguratio
 import dev.turingcomplete.intellijdevelopertoolsplugin.DeveloperToolConfiguration.PropertyType.SECRET
 import dev.turingcomplete.intellijdevelopertoolsplugin._internal.common.ValueProperty
 import dev.turingcomplete.intellijdevelopertoolsplugin._internal.common.ValueProperty.Companion.RESET_CHANGE_ID
-import dev.turingcomplete.intellijdevelopertoolsplugin._internal.common.uncheckedCastTo
 import dev.turingcomplete.intellijdevelopertoolsplugin._internal.settings.DeveloperToolsApplicationSettings
 import dev.turingcomplete.intellijdevelopertoolsplugin._internal.settings.DeveloperToolsInstanceSettings.Companion.assertPersistableType
 import java.util.*
 import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.CopyOnWriteArrayList
+import kotlin.reflect.safeCast
 
 class DeveloperToolConfiguration(
   var name: String,
@@ -27,6 +27,7 @@ class DeveloperToolConfiguration(
   // In this case, the `persistentProperties` need to be persisted again.
   internal var wasConsumedByDeveloperTool = false
   private val changeListeners = CopyOnWriteArrayList<ChangeListener>()
+  private val resetListeners = CopyOnWriteArrayList<ResetListener>()
   var isResetting = false
     internal set
 
@@ -50,6 +51,15 @@ class DeveloperToolConfiguration(
     changeListeners.remove(changeListener)
   }
 
+  fun addResetListener(parentDisposable: Disposable, resetListener: ResetListener) {
+    resetListeners.add(resetListener)
+    Disposer.register(parentDisposable) { resetListeners.remove(resetListener) }
+  }
+
+  fun removeResetListener(resetListener: ResetListener) {
+    resetListeners.remove(resetListener)
+  }
+
   fun reset(
     type: PropertyType? = null,
     loadExamples: Boolean = DeveloperToolsApplicationSettings.instance.loadExamples
@@ -61,6 +71,7 @@ class DeveloperToolConfiguration(
           property.reset(loadExamples)
           fireConfigurationChanged(property.reference)
         }
+      resetListeners.forEach { it.configurationReset() }
     }
     finally {
       isResetting = false
@@ -89,7 +100,7 @@ class DeveloperToolConfiguration(
   ): ValueProperty<T> {
     val type = assertPersistableType(defaultValue::class, propertyType)
     val existingPropertyValue = persistentProperties[key]?.value
-    val initialValue: T = existingPropertyValue?.uncheckedCastTo(type) ?: let {
+    val initialValue: T = type.safeCast(existingPropertyValue) ?: let {
       if (DeveloperToolsApplicationSettings.instance.loadExamples && example != null) example else defaultValue
     }
     val valueProperty = ValueProperty(initialValue).apply {
@@ -150,9 +161,18 @@ class DeveloperToolConfiguration(
 
   // -- Inner Type -------------------------------------------------------------------------------------------------- //
 
-  interface ChangeListener {
+  @FunctionalInterface
+  fun interface ChangeListener {
 
     fun configurationChanged(property: ValueProperty<out Any>)
+  }
+
+  // -- Inner Type -------------------------------------------------------------------------------------------------- //
+
+  @FunctionalInterface
+  fun interface ResetListener {
+
+    fun configurationReset()
   }
 
   // -- Inner Type -------------------------------------------------------------------------------------------------- //