+++ CHANGELOG.md
@@ -3,14 +3,22 @@
 ## Unreleased
 
 ### Added
-- Add time conversion tool
-- Add 'What's New' overview to the main menu
 
 ### Changed
 
 ### Removed
 
 ### Fixed
+
+## 3.3.0 - 2024-01-15
+
+### Added
+
+- Add time conversion tool
+- Add 'What's New' overview to the main menu
+
+### Fixed
+
 - Inputs and configurations for tools that were not opened were lost the next time the settings were saved.
 - Loading the tool window settings causes an error if one of the tools has a saved secret.
 - In the JWT Encoder/Decoder the "Public Key" field was visible for the HMAC signature algorithm.